✨ remove deprecated code from main codebase
