+++ .env.example-full
@@ -23,6 +23,7 @@ GDOCS_ARTICLE_DUPLICATION_TEMPLATE_ID=
 GDOCS_DATA_INSIGHT_DUPLICATION_TEMPLATE_ID=
 # Template document used as the starting point for an automation workflow
 GDOCS_DATA_INSIGHT_API_TEMPLATE_ID=
+# TODO: remove this once https://github.com/owid/owid-grapher/pull/4852 is merged
 GDOCS_DETAILS_ON_DEMAND_ID=
 GDOCS_DONATE_FAQS_DOCUMENT_ID=
 GDOCS_BACKPORTING_TARGET_FOLDER=  # optional
