+++ packages/@ourworldindata/types/src/index.ts
@@ -337,14 +337,11 @@ export {
     type OwidGdocHomepageMetadata,
     DATA_INSIGHTS_INDEX_PAGE_SIZE,
     type OwidGdoc,
-    type RawDetail,
     OwidGdocType,
     type OwidGdocStickyNavItem,
     type OwidGdocJSON,
     type FaqDictionary,
     type EnrichedDetail,
-    type DEPRECATED_DetailDictionary,
-    type DEPRECATED_EnrichedDetail,
     type EnrichedFaq,
     type DetailDictionary,
     GdocsContentSource,
